Most of the asset backend now working(not allocate/deallocate)
UI still needs fixing, however through the console logs we can see that the data is being collected from the DB correctly. Dashboard displays all data correctly, however the pages that use a table are only able to display last pinged time. This is likely a UI issue from the change to a new DB but I can't seem to find the exact cause